backend/requirements.txt
langgraph>=1.0.0
langchain>=1.2.0
langchain-openai>=1.1.0
langchain-community>=0.4.0
openai>=2.0.0
tavily-python>=0.5.0
duckduckgo-search>=8.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
rich>=14.0.0
beautifulsoup4>=4.14.0
aiohttp>=3.9.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
